+++ R/Sobolev.R
@@ -206,8 +206,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       # Halve K_max since we compute the odd/even coefficients separately
       K_max <- K_max %/% 2
 
-      # Sequence of indexes
-      k <- 1:K_max
+      # Sequence of indexes (seq_len is empty, not c(1, 0), when K_max < 2)
+      k <- seq_len(K_max)
 
       # log(v_{2 * k - 1}^2)
       log_v2km12 <- log(2 / pi) - log((2 * k - 1)^2)
@@ -250,8 +250,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       # Halve K_max since the even coefficients are zero
       K_max <- K_max %/% 2
 
-      # Sequence of indexes
-      k <- 1:K_max
+      # Sequence of indexes (seq_len is empty, not c(1, 0), when K_max < 2)
+      k <- seq_len(K_max)
 
       # log(v_{2 * k - 1}^2)
       log_v2km12 <- (p - 2) * log(2) + lgamma(alpha + 1) + lgamma(k + alpha) +
@@ -262,7 +262,7 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       log_v2km12 <- c(rbind(log_v2km12, -Inf))
 
       # log(d_{p, k})
-      log_dk <- d_p_k(p = p, k = 1:(2 * K_max), log = TRUE)
+      log_dk <- d_p_k(p = p, k = seq_len(2 * K_max), log = TRUE)
 
       # Log weights and dfs
       log_weights <- log_v2km12
@@ -273,8 +273,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       # Halve K_max since the odd coefficients are zero
       K_max <- K_max %/% 2
 
-      # Sequence of indexes
-      k <- 1:K_max
+      # Sequence of indexes (seq_len is empty, not c(1, 0), when K_max < 2)
+      k <- seq_len(K_max)
 
       # log(v_{2 * k}^2)
       log_v2k2 <- log((p - 1) * (2 * k - 1) / (8 * pi * (2 * k + p - 1))) +
@@ -284,7 +284,7 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       log_v2k2 <- c(rbind(-Inf, log_v2k2))
 
       # log(d_{p, k})
-      log_dk <- d_p_k(p = p, k = 1:(2 * K_max), log = TRUE)
+      log_dk <- d_p_k(p = p, k = seq_len(2 * K_max), log = TRUE)
 
       # Log weights and dfs
       log_weights <- log_v2k2
@@ -295,8 +295,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       # Halve K_max since we sum 4 * Ajne and Gine_Gn
       K_max <- K_max %/% 2
 
-      # Sequence of indexes
-      k <- 1:K_max
+      # Sequence of indexes (seq_len is empty, not c(1, 0), when K_max < 2)
+      k <- seq_len(K_max)
 
       # log(v_{2 * k - 1}^2)
       log_v2km12 <- (p - 2) * log(2) + lgamma(alpha + 1) + lgamma(k + alpha) +
@@ -308,7 +308,7 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
         2 * (lgamma(alpha + 0.5) + lgamma(k - 0.5) - lgamma(k + alpha + 0.5))
 
       # log(d_{p, k})
-      log_dk <- d_p_k(p = p, k = 1:(2 * K_max), log = TRUE)
+      log_dk <- d_p_k(p = p, k = seq_len(2 * K_max), log = TRUE)
 
       # Log weights and dfs
       log_weights <- c(rbind(log(4) + log_v2km12, log_v2k2))
@@ -550,8 +550,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       # Halve K_max since we compute the odd/even coefficients separately
       K_max <- K_max %/% 2
 
-      # Sequence of indexes
-      k <- 1:K_max
+      # Sequence of indexes (seq_len is empty, not c(1, 0), when K_max < 2)
+      k <- seq_len(K_max)
 
       # log(v_{2 * k - 1}^2)
       log_v2km12 <- log(2 * k - 1) + log(4 * (k - 1) + p) - log(2 * k + p - 3) +
@@ -571,7 +571,7 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
       log_vk2 <- bk_to_vk2(bk = log_vk2, p = p, log = TRUE)
 
       # log(d_{p, k})
-      log_dk <- d_p_k(p = p, k = 1:(2 * K_max), log = TRUE)
+      log_dk <- d_p_k(p = p, k = seq_len(2 * K_max), log = TRUE)
 
       # Log weights and dfs
       log_weights <- log_vk2
@@ -640,7 +640,8 @@ weights_dfs_Sobolev <- function(p, K_max = 1e3, thre = 1e-3, type,
 d_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE")[1], K_max = 1e3,
                       thre = 1e-3, Rothman_t = 1 / 3, Pycke_q = 0.5,
                       Riesz_s = 1, Poisson_rho = 0.5, Softmax_kappa = 1,
-                      Stereo_a = 0, Sobolev_vk2 = c(0, 0, 1), ncps = 0,
+                      Stereo_a = 0, Stein_cf = FALSE,
+                      Sobolev_vk2 = c(0, 0, 1), ncps = 0,
                       verbose = TRUE, N = 320, x_tail = NULL, ...) {
 
   weights_dfs <- weights_dfs_Sobolev(p = p, K_max = K_max, thre = thre,
@@ -648,7 +649,7 @@ d_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE")[1], K_max = 1e3,
                                      Pycke_q = Pycke_q, Riesz_s = Riesz_s,
                                      Poisson_rho = Poisson_rho,
                                      Softmax_kappa = Softmax_kappa,
-                                     Stereo_a = Stereo_a,
+                                     Stereo_a = Stereo_a, Stein_cf = Stein_cf,
                                      Sobolev_vk2 = Sobolev_vk2,
                                      verbose = verbose, Gauss = TRUE, N = N,
                                      x_tail = x_tail)
@@ -663,7 +664,8 @@ d_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE")[1], K_max = 1e3,
 p_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE", "MC")[1],
                       K_max = 1e3, thre = 1e-3, Rothman_t = 1 / 3,
                       Pycke_q = 0.5, Riesz_s = 1, Poisson_rho = 0.5,
-                      Softmax_kappa = 1, Stereo_a = 0, Sobolev_vk2 = c(0, 0, 1),
+                      Softmax_kappa = 1, Stereo_a = 0, Stein_cf = FALSE,
+                      Sobolev_vk2 = c(0, 0, 1),
                       ncps = 0, verbose = TRUE, N = 320, x_tail = NULL, ...) {
 
   weights_dfs <- weights_dfs_Sobolev(p = p, K_max = K_max, thre = thre,
@@ -671,7 +673,7 @@ p_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE", "MC")[1],
                                      Pycke_q = Pycke_q, Riesz_s = Riesz_s,
                                      Poisson_rho = Poisson_rho,
                                      Softmax_kappa = Softmax_kappa,
-                                     Stereo_a = Stereo_a,
+                                     Stereo_a = Stereo_a, Stein_cf = Stein_cf,
                                      Sobolev_vk2 = Sobolev_vk2,
                                      verbose = verbose, Gauss = TRUE, N = N,
                                      x_tail = x_tail)
@@ -686,7 +688,8 @@ p_Sobolev <- function(x, p, type, method = c("I", "SW", "HBE", "MC")[1],
 q_Sobolev <- function(u, p, type, method = c("I", "SW", "HBE", "MC")[1],
                       K_max = 1e3, thre = 1e-3, Rothman_t = 1 / 3,
                       Pycke_q = 0.5, Riesz_s = 1, Poisson_rho = 0.5,
-                      Softmax_kappa = 1, Stereo_a = 0, Sobolev_vk2 = c(0, 0, 1),
+                      Softmax_kappa = 1, Stereo_a = 0, Stein_cf = FALSE,
+                      Sobolev_vk2 = c(0, 0, 1),
                       ncps = 0, verbose = TRUE, N = 320, x_tail = NULL, ...) {
 
   weights_dfs <- weights_dfs_Sobolev(p = p, K_max = K_max, thre = thre,
@@ -694,7 +697,7 @@ q_Sobolev <- function(u, p, type, method = c("I", "SW", "HBE", "MC")[1],
                                      Pycke_q = Pycke_q, Riesz_s = Riesz_s,
                                      Poisson_rho = Poisson_rho,
                                      Softmax_kappa = Softmax_kappa,
-                                     Stereo_a = Stereo_a,
+                                     Stereo_a = Stereo_a, Stein_cf = Stein_cf,
                                      Sobolev_vk2 = Sobolev_vk2,
                                      verbose = verbose, Gauss = TRUE, N = N,
                                      x_tail = x_tail)